requirements.txt
altair==4.2.0
attrs==22.2.0
blinker==1.5
cachetools==5.2.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
commonmark==0.9.1
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
entrypoints==0.4
fonttools==4.38.0
gitdb==4.0.10
GitPython==3.1.30
idna==3.4
importlib-metadata==6.0.0
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.2
numpy==1.23.4
packaging==23.0
pandas==1.5.1
Pillow==9.3.0
pip==22.3.1
plotly==5.11.0
protobuf==3.20.3
pyarrow==10.0.1
pydeck==0.8.0
Pygments==2.14.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
rich==13.0.1
scikit-learn==1.4.0
scipy==1.10.0
seaborn==0.12.2
semver==2.13.0
setuptools==65.6.3
six==1.16.0
smmap==5.0.0
streamlit==1.14.1
tenacity==8.1.0
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
validators==0.20.0
watchdog==2.2.1
wheel==0.37.1
zipp==3.11.0
